Delete unused import and templates/sitemap.xml
