+++ story/sitemap.py
@@ -9,5 +9,13 @@ class BlogSiteMap(Sitemap):
     def items(self):
         return Story.objects.all()
 
+    def location(self, item):
+        url = item.post_url
+        return url.replace('https://', '')
+
     def lastmod(self, obj):
         return obj.created
+
+    def _urls(self, page, protocol, domain):
+        return super(BlogSiteMap, self)._urls(
+            page=page, protocol='https', domain='')
