+++ tests/story/conftest.py
@@ -1,7 +1,5 @@
 import pytest
 
-from story.models import Story
-
 
 @pytest.fixture
 def data_dict():
@@ -22,16 +20,3 @@ def data_dict():
         'is_story': True,
         'post_url': 'story-1-url',
     }
-
-
-@pytest.fixture
-def blog_posts(db):
-    return Story.objects.bulk_create(
-        [
-            Story(name='Post 1', content='Lorem ipsum dolor sit amet', is_story=False, post_url='post-1-url'),
-            Story(name='Post 2', content='Lorem ipsum dolor sit amet', is_story=False, post_url='post-2-url'),
-            Story(name='Post 3', content='Lorem ipsum dolor sit amet', is_story=False, post_url='post-3-url'),
-            Story(name='Post 4', content='Lorem ipsum dolor sit amet', is_story=False, post_url='post-4-url'),
-            Story(name='Post 5', content='Lorem ipsum dolor sit amet', is_story=False, post_url='post-5-url')
-        ]
-    )
